move entities from models folder to entities folder fixing namespace
