--- OngProject/Core/Models/RoleModel.cs
@@ -1,20 +1,0 @@
-﻿using OngProject.Entities;
-using System;
-using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace OngProject.Core.Models
-{
-    public class RoleModel : EntityBase
-    {
-      
-        [Required]
-       
-        public string NameRole { get; set; }
-        public string DescriptionRole { get; set; }
-
-
-    }
-}
